server: Get rid of former - bad - benefit file sorting middleware
